Foreign.Storable.Complex

Description

This module provides a Storable instance for Complex which is binary compatible with C99, C++ and Fortran complex data types. It's only purpose is to provide a standard location for this instance so that other packages needing this instance can play nicely together.